Fix the typo of (a) new Node(NodeIDType.OPENFLOW) to new Node(NodeIDType.SNMP) (b) NodeConnectorIDType("SNMP", Long.class, "SNMP") to ("SNMP", Short.class, "SNMP")

+++ snmp4sdn/src/test/java/org/opendaylight/snmp4sdn/internal/DiscoveryServiceTest.java
@@ -47,7 +47,7 @@ public class DiscoveryServiceTest {
 
     public DiscoveryServiceTest(){
         Node.NodeIDType.registerIDType("SNMP", Long.class);
-        NodeConnector.NodeConnectorIDType.registerIDType("SNMP", Long.class, "SNMP");
+        NodeConnector.NodeConnectorIDType.registerIDType("SNMP", Short.class, "SNMP");
         /*
         try {
             testDiscoveryService();
